requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.9
django-allauth==65.13.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-simpleui==2025.6.24
idna==3.11
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
